pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-council"
version = "0.1.0"
description = "Lightweight read-only multi-agent council for coding projects."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "httpx>=0.27.0",
  "mcp>=1.0.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
llm-council = "llm_council.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["llm_council*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
